backend/requirements.txt
aioredis
bcrypt
beanie
black
iso4217parse # for currency codes
itsdangerous # for starlette sessions
Faker
fastapi
fastapi-jwt-auth
fastapi-mail
httpx
python-decouple
uvicorn